added Router to app.js and moved classes to pages
