+++ client/src/App.css
@@ -1,3 +1,8 @@
 div, button {
   margin: 10px;
 }
+
+.wrapper {
+  min-height: 100%;
+  padding-bottom: 100px;
+}
